Sass update and development the static css for the mobile version.
Creating scss structure, and working on the css and the html code.
Finalizing the first mobile version of the css code for the home, and the menu pages.
Html changes on the component responsible for the items selection in the menu page for better combability with the future js functionality.